compile wasm fixes
