+++ rust/poker_eval/Cargo.toml
@@ -47,6 +47,7 @@ enum_dispatch = "0.3.12"
 
 num-integer = "0.1"
 
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 jammdb = "0.11"
 redb = "1.4"
 rmp-serde = "1.1.2"
