+++ rust/poker_eval/src/core/mod.rs
@@ -11,3 +11,6 @@ pub use card_util::*;
 
 mod test_util;
 pub use test_util::*;
+
+mod deck;
+pub use deck::*;
