+++ rust/poker_eval/src/eval/board_eval_cache_redb.rs
@@ -1,7 +1,7 @@
 use redb::{Database, Error as ReDbError, ReadableTable, TableDefinition, ReadTransaction};
 use serde::{Serialize, de::DeserializeOwned};
 
-use crate::{CombinatorialIndex, Card,};
+use crate::{CombinatorialIndex, Card, BoardTexture, calc_board_texture,};
 
 //u32 is usually  enough
 //In the worst case we have 5 cards * 2 cards
@@ -108,4 +108,99 @@ where P : ProduceEvalResult<R>, R :  Serialize + DeserializeOwned,
         write_txn.commit()?;
         Ok(())
     }
+}
+
+
+pub struct ProduceFlopTexture {
+
+}
+
+impl ProduceFlopTexture {
+    pub fn new() -> Self {
+        ProduceFlopTexture {
+
+        }
+    }
+}
+
+impl ProduceEvalResult<BoardTexture> for ProduceFlopTexture {
+    fn produce_eval_result(cards: &[Card]) -> BoardTexture {
+        calc_board_texture(cards)
+    }
+
+}
+
+#[cfg(test)]
+mod tests {
+
+    use std::time::Instant;
+
+    use log::{info, debug};
+
+    use crate::{init_test_logger, Deck, CardVec, board_eval_cache_redb::{EvalCacheReDb, FLOP_TEXTURE_PATH}};
+
+    use super::*;
+
+    // cargo test cache_perf --lib --release -- --nocapture
+
+    //a bit slow
+    #[test]
+    fn test_cache_perf() {
+        init_test_logger();
+
+        let mut agent_deck = Deck::new();
+        let mut cards: Vec<Card> = Vec::new();
+       
+        cards.clear();
+        agent_deck.reset();
+        //delete if exists
+        //std::fs::remove_file(db_name).unwrap_or_default();
+
+        //let mut flop_texture_db = FlopTextureJamDb::new(db_name).unwrap();
+
+        
+        //let mut flop_texture_db = FlopTextureReDb::new(re_db_name).unwrap();
+        let mut flop_texture_db: EvalCacheReDb<ProduceFlopTexture, _> =
+             EvalCacheReDb::new(FLOP_TEXTURE_PATH).unwrap();
+        let now = Instant::now();
+        let iter_count = 100_000;
+        // Code block to measure.
+        {
+            for i in 0..iter_count {
+                cards.push(agent_deck.get_unused_card().unwrap().try_into().unwrap());
+                cards.push(agent_deck.get_unused_card().unwrap().try_into().unwrap());
+                cards.push(agent_deck.get_unused_card().unwrap().try_into().unwrap());
+                let _texture = flop_texture_db.get_put(&cards).unwrap();
+                agent_deck.clear_used_card(cards[0]);
+                agent_deck.clear_used_card(cards[1]);
+                agent_deck.clear_used_card(cards[2]);
+                cards.pop();
+                cards.pop();
+                cards.pop();
+
+                if flop_texture_db.cache_misses > 0 && flop_texture_db.cache_misses % 1000 == 0 {
+                    println!("Iter {}", i);
+                    info!(
+                        "Cache hits {} misses {}",
+                        flop_texture_db.cache_hits, flop_texture_db.cache_misses
+                    );
+                }
+                if flop_texture_db.cache_hits > 0 && flop_texture_db.cache_hits % 100_000 == 0 {
+                    println!("Iter {}", i);
+                    info!(
+                        "Cache hits {} misses {}",
+                        flop_texture_db.cache_hits, flop_texture_db.cache_misses
+                    );
+                }
+            }
+        }
+
+        let elapsed = now.elapsed();
+        println!("Elapsed: {:.2?}", elapsed);
+        info!(
+            "Cache hits {} misses {}",
+            flop_texture_db.cache_hits, flop_texture_db.cache_misses
+        );
+    }
+
 }
