+++ rust/poker_eval/src/eval/flop_texture.rs
@@ -39,7 +39,7 @@ use serde::{Deserialize, Serialize};
 // use rmps crate to serialize structs using the MessagePack format
 use crate::{
     calc_cards_metrics, partial_rank_cards, rank_cards, Card, CardValue, 
-    HoleCards, StraightDrawType, CardVec, board_eval_cache_redb::ProduceEvalResult,
+    HoleCards, StraightDrawType, CardVec, 
 };
 
 
@@ -206,99 +206,19 @@ pub fn calc_board_texture(cards: &[Card]) -> BoardTexture {
     texture
 }
 
-pub struct ProduceFlopTexture {
-
-}
-
-impl ProduceFlopTexture {
-    pub fn new() -> Self {
-        ProduceFlopTexture {
-
-        }
-    }
-}
-
-impl ProduceEvalResult<BoardTexture> for ProduceFlopTexture {
-    fn produce_eval_result(cards: &[Card]) -> BoardTexture {
-        calc_board_texture(cards)
-    }
-
-}
 
 #[cfg(test)]
 mod tests {
 
-    use std::time::Instant;
-
     use log::{info, debug};
 
-    use crate::{init_test_logger, AgentDeck, CardVec, board_eval_cache_redb::{EvalCacheReDb, FLOP_TEXTURE_PATH}};
+    use crate::{init_test_logger, CardVec};
 
     use super::*;
 
-    // cargo test cache_perf --lib --release -- --nocapture
+    
 
-    //a bit slow
-    #[test]
-    //#[allow(dead_code)]
-    fn test_cache_perf() {
-        init_test_logger();
-
-        let mut agent_deck = AgentDeck::new();
-        let mut cards: Vec<Card> = Vec::new();
-       
-        cards.clear();
-        agent_deck.reset();
-        //delete if exists
-        //std::fs::remove_file(db_name).unwrap_or_default();
-
-        //let mut flop_texture_db = FlopTextureJamDb::new(db_name).unwrap();
-
-        
-        //let mut flop_texture_db = FlopTextureReDb::new(re_db_name).unwrap();
-        let mut flop_texture_db: EvalCacheReDb<ProduceFlopTexture, _> =
-             EvalCacheReDb::new(FLOP_TEXTURE_PATH).unwrap();
-        let now = Instant::now();
-        let iter_count = 100_000;
-        // Code block to measure.
-        {
-            for i in 0..iter_count {
-                cards.push(agent_deck.get_unused_card().unwrap().try_into().unwrap());
-                cards.push(agent_deck.get_unused_card().unwrap().try_into().unwrap());
-                cards.push(agent_deck.get_unused_card().unwrap().try_into().unwrap());
-                let _texture = flop_texture_db.get_put(&cards).unwrap();
-                agent_deck.clear_used_card(cards[0]);
-                agent_deck.clear_used_card(cards[1]);
-                agent_deck.clear_used_card(cards[2]);
-                cards.pop();
-                cards.pop();
-                cards.pop();
-
-                if flop_texture_db.cache_misses > 0 && flop_texture_db.cache_misses % 1000 == 0 {
-                    println!("Iter {}", i);
-                    info!(
-                        "Cache hits {} misses {}",
-                        flop_texture_db.cache_hits, flop_texture_db.cache_misses
-                    );
-                }
-                if flop_texture_db.cache_hits > 0 && flop_texture_db.cache_hits % 100_000 == 0 {
-                    println!("Iter {}", i);
-                    info!(
-                        "Cache hits {} misses {}",
-                        flop_texture_db.cache_hits, flop_texture_db.cache_misses
-                    );
-                }
-            }
-        }
-
-        let elapsed = now.elapsed();
-        println!("Elapsed: {:.2?}", elapsed);
-        info!(
-            "Cache hits {} misses {}",
-            flop_texture_db.cache_hits, flop_texture_db.cache_misses
-        );
-    }
-
+    
     #[test]
     fn test_board_texture() {
         init_test_logger();
