+++ rust/poker_eval/src/eval/partial_rank.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::{
     calc_bitset_cards_metrics, count_higher, count_lower, rank_straight, value_set_iterator,
-    BitSetCardsMetrics, Card, CardValue, HoleCards, ValueSetType, board_hc_eval_cache_redb::ProduceEvalWithHcResult,
+    BitSetCardsMetrics, Card, CardValue, HoleCards, ValueSetType,
 };
 
 //for pairs, 2 pair, sets, quads, full houses
@@ -613,17 +613,6 @@ pub fn partial_rank_cards(hole_cards: &HoleCards, board: &[Card]) -> PartialRank
     partial_ranks
 }
 
-pub struct ProducePartialRankCards {
-
-}
-
-impl ProduceEvalWithHcResult<PartialRankContainer> for ProducePartialRankCards {
-    
-    fn produce_eval_result(board: &[Card], hole_cards: &HoleCards, ) -> PartialRankContainer {
-        partial_rank_cards(&hole_cards, board)
-    }
-}
-
 #[cfg(test)]
 mod tests {
 
@@ -631,7 +620,7 @@ mod tests {
 
     use log::info;
 
-    use crate::{CardVec, AgentDeck, init_test_logger, board_hc_eval_cache_redb::{EvalCacheWithHcReDb, PARTIAL_RANK_PATH}};
+    use crate::{CardVec, Deck, init_test_logger, board_hc_eval_cache_redb::{EvalCacheWithHcReDb, PARTIAL_RANK_PATH}};
 
     use super::*;
 
@@ -1102,67 +1091,4 @@ mod tests {
         );
     }
 
-    #[test]
-    fn test_cache_partial_rank() {
-        init_test_logger();
-
-        let mut agent_deck = AgentDeck::new();
-        let mut cards: Vec<Card> = Vec::new();
-        
-        cards.clear();
-        agent_deck.reset();
-        //delete if exists
-        //std::fs::remove_file(db_name).unwrap_or_default();
-
-        //let mut flop_texture_db = FlopTextureJamDb::new(db_name).unwrap();
-
-        
-        //let mut flop_texture_db = FlopTextureReDb::new(re_db_name).unwrap();
-        let mut partial_rank_db: EvalCacheWithHcReDb<ProducePartialRankCards, _> =
-        EvalCacheWithHcReDb::new(PARTIAL_RANK_PATH).unwrap();
-        let now = Instant::now();
-        let iter_count = 500_000;
-        // Code block to measure.
-        {
-            for i in 0..iter_count {
-                cards.push(agent_deck.get_unused_card().unwrap().try_into().unwrap());
-                cards.push(agent_deck.get_unused_card().unwrap().try_into().unwrap());
-                cards.push(agent_deck.get_unused_card().unwrap().try_into().unwrap());
-                let hole1: Card = agent_deck.get_unused_card().unwrap().try_into().unwrap();
-                let hole2: Card = agent_deck.get_unused_card().unwrap().try_into().unwrap();
-                let hole_cards: HoleCards = HoleCards::new(hole1, hole2).unwrap();
-                let _texture = partial_rank_db.get_put(&cards, &hole_cards).unwrap();
-                agent_deck.clear_used_card(cards[0]);
-                agent_deck.clear_used_card(cards[1]);
-                agent_deck.clear_used_card(cards[2]);
-                cards.pop();
-                cards.pop();
-                cards.pop();
-                agent_deck.clear_used_card(hole1);
-                agent_deck.clear_used_card(hole2);
-
-                if partial_rank_db.cache_misses > 0 && partial_rank_db.cache_misses % 1000 == 0 {
-                    println!("Iter {}", i);
-                    info!(
-                        "Cache hits {} misses {}",
-                        partial_rank_db.cache_hits, partial_rank_db.cache_misses
-                    );
-                }
-                if partial_rank_db.cache_hits > 0 && partial_rank_db.cache_hits % 100_000 == 0 {
-                    println!("Iter {}", i);
-                    info!(
-                        "Cache hits {} misses {}",
-                        partial_rank_db.cache_hits, partial_rank_db.cache_misses
-                    );
-                }
-            }
-        }
-
-        let elapsed = now.elapsed();
-        println!("Elapsed: {:.2?}", elapsed);
-        info!(
-            "Cache hits {} misses {}",
-            partial_rank_db.cache_hits, partial_rank_db.cache_misses
-        );
-    }
 }
