+++ rust/poker_eval/src/game/agents/agent_trait.rs
@@ -1,9 +1,6 @@
-use crate::{ActionEnum, GameState, HoleCards, PlayerState};
+use crate::{GameState, HoleCards, PlayerState, CommentedAction};
 
-pub struct AgentDecision {
-    pub action: ActionEnum,
-    pub comment: Option<String>,
-}
+
 
 pub trait Agent {
     //Get hand cards with index_to_card_pair
@@ -12,7 +9,7 @@ pub trait Agent {
         &mut self,
         player_state: &PlayerState,
         game_state: &GameState,
-    ) -> AgentDecision;
+    ) -> CommentedAction;
 
     fn get_hole_cards(&self) -> HoleCards;
