+++ rust/poker_eval/src/game/agents/agent_util.rs
@@ -1,83 +1,20 @@
-use log::warn;
-use rand::{rngs::StdRng, Rng, SeedableRng};
-
-use crate::{Card, CardUsedType, HoleCards, InitialPlayerState, Position};
+use crate::{Card, HoleCards, InitialPlayerState, Position, Deck};
 
 use super::Agent;
 
-pub struct AgentDeck {
-    rng: StdRng,
-    used_cards: CardUsedType, //= CardUsedType::default();
-}
+pub fn set_agent_hole_cards(deck: &mut Deck, agents: &mut Vec<Box<dyn Agent>>) {
+    for agent_index in 0..agents.len() {
+        let agent = &mut agents[agent_index];
 
-const MAX_RAND_NUMBER_ATTEMPS: usize = 10_000;
+        let card1 = deck.get_unused_card().unwrap();
 
-impl AgentDeck {
-    pub fn new() -> Self {
-        let rng = StdRng::seed_from_u64(42);
-        AgentDeck {
-            rng,
-            used_cards: CardUsedType::default(),
-        }
-    }
-
-    pub fn reset(&mut self) {
-        self.used_cards = CardUsedType::default();
-    }
-
-    pub fn set_agent_hole_cards(&mut self, agents: &mut Vec<Box<dyn Agent>>) {
-        for agent_index in 0..agents.len() {
-            let agent = &mut agents[agent_index];
-
-            let card1 = self.get_unused_card().unwrap();
-
-            let card2 = self.get_unused_card().unwrap();
-            let agent_hole_cards = HoleCards::new(
-                Card::try_from(card1).unwrap(),
-                Card::try_from(card2).unwrap(),
-            )
-            .unwrap();
-            agent.set_hole_cards(agent_hole_cards);
-        }
-    }
-
-    pub fn get_board(&mut self) -> Vec<Card> {
-        let mut board = Vec::new();
-        for _ in 0..5 {
-            let card = self.get_unused_card().unwrap();
-            board.push(Card::try_from(card).unwrap());
-        }
-        board
-    }
-
-    pub fn clear_used_card(&mut self, card: Card) {
-        let card_index: usize = card.into();
-        let count_before = self.used_cards.count_ones();
-        self.used_cards.set(card_index, false);
-        let count_after = self.used_cards.count_ones();
-        assert_eq!(count_after + 1, count_before);
-    }
-
-    pub fn get_unused_card(&mut self) -> Option<usize> {
-        let mut attempts = 0;
-        loop {
-            let rand_int: usize = self.rng.gen_range(0..52);
-            assert!(rand_int < 52);
-            //let card = Card::from(rand_int);
-            if !self.used_cards[rand_int] {
-                self.used_cards.set(rand_int, true);
-                return Some(rand_int);
-            }
-            attempts += 1;
-            if attempts > MAX_RAND_NUMBER_ATTEMPS {
-                warn!(
-                    "Unable to find unused card after {} attempts, # of used cards {}",
-                    attempts,
-                    self.used_cards.count_ones()
-                );
-                return None;
-            }
-        }
+        let card2 = deck.get_unused_card().unwrap();
+        let agent_hole_cards = HoleCards::new(
+            Card::try_from(card1).unwrap(),
+            Card::try_from(card2).unwrap(),
+        )
+        .unwrap();
+        agent.set_hole_cards(agent_hole_cards);
     }
 }
