+++ rust/poker_eval/src/game/agents/mod.rs
@@ -3,9 +3,11 @@ mod agent_util;
 mod just_fold;
 mod passive_calling_station;
 mod tag;
+mod agent_source;
 
 pub use agent_trait::*;
 pub use agent_util::*;
 pub use just_fold::*;
 pub use passive_calling_station::*;
 pub use tag::*;
+pub use agent_source::*;
