+++ rust/poker_eval/src/game/game_runner_source.rs
@@ -1,16 +1,17 @@
 use crate::{
     game::game_log_source::GameLogSource, Card, ChipType, GameState, HoleCards, InitialPlayerState,
-    PlayerState, PokerError,
+    PlayerState, PokerError, CommentedAction,
 };
 use enum_dispatch::enum_dispatch;
 
-use crate::game::agent_source::AgentSource;
+#[cfg(not(target_arch = "wasm32"))]
+use crate::game::agents::AgentSource;
 
-use super::agents::AgentDecision;
 
 #[enum_dispatch]
 pub enum GameRunnerSourceEnum {
     GameLogSource,
+    #[cfg(not(target_arch = "wasm32"))]
     AgentSource,
 }
 
@@ -25,7 +26,7 @@ pub trait GameRunnerSource {
         &mut self,
         player_state: &PlayerState,
         game_state: &GameState,
-    ) -> Result<AgentDecision, PokerError>;
+    ) -> Result<CommentedAction, PokerError>;
 
     //get cards for player?
     fn get_hole_cards(&self, player_index: usize) -> Result<HoleCards, PokerError>;
